+++ src/integration-helpers.ts
@@ -52,3 +52,22 @@ export const makeEventFnSchema = (
     variants as unknown as [z.ZodTypeAny, z.ZodTypeAny, ...z.ZodTypeAny[]],
   );
 };
+/**
+ * @desc isNullable() and isOptional() validate the schema's input
+ * @desc They always return true in case of coercion, which should be taken into account when depicting response
+ */
+export const hasCoercion = (schema: z.ZodTypeAny): boolean =>
+  "coerce" in schema._def && typeof schema._def.coerce === "boolean"
+    ? schema._def.coerce
+    : false;
+
+export const tryToTransform = <T>(
+  schema: z.ZodEffects<z.ZodTypeAny, T>,
+  sample: T,
+) => {
+  try {
+    return typeof schema.parse(sample);
+  } catch (e) {
+    return undefined;
+  }
+};
