+++ src/zts.ts
@@ -1,7 +1,12 @@
 import ts from "typescript";
 import { z } from "zod";
 
-import { lcFirst, makeCleanId } from "./integration-helpers";
+import {
+  hasCoercion,
+  lcFirst,
+  makeCleanId,
+  tryToTransform,
+} from "./integration-helpers";
 import { HandlingRules, walkSchema } from "./schema-walker";
 import {
   LiteralType,
@@ -23,26 +28,6 @@ const samples = {
   [ts.SyntaxKind.UndefinedKeyword]: undefined,
   [ts.SyntaxKind.UnknownKeyword]: undefined,
 } satisfies Partial<Record<ts.KeywordTypeSyntaxKind, unknown>>;
-
-/**
- * @desc isNullable() and isOptional() validate the schema's input
- * @desc They always return true in case of coercion, which should be taken into account when depicting response
- */
-export const hasCoercion = (schema: z.ZodTypeAny): boolean =>
-  "coerce" in schema._def && typeof schema._def.coerce === "boolean"
-    ? schema._def.coerce
-    : false;
-
-export const tryToTransform = <T>(
-  schema: z.ZodEffects<z.ZodTypeAny, T>,
-  sample: T,
-) => {
-  try {
-    return typeof schema.parse(sample);
-  } catch (e) {
-    return undefined;
-  }
-};
 
 const onLiteral: Producer<z.ZodLiteral<LiteralType>> = ({
   schema: { value },
